Perform a final cleanup of the optimizer rules tests.
